update sliding windows
update sliding windows.
these are to ensure a correct result when comparing to groundtruths.
because the comparison is based on JI and sidecars are labelled tightly
to their bounds which have proportions that are unlike the square
collations returned from opencv, we use these windows to better ensure
proper result reporting.